requirements.txt
# requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2

